requirements.txt
Django==4.0.5
djangorestframework==3.13.1
requests==2.28.1
django-cors-headers==3.13.0
django-heroku==0.3.1
gunicorn==20.0.4
whitenoise==5.2.0
python-dotenv==0.15.0
psycopg2-binary
django-storages
WeasyPrint